requirements.txt
fastapi
uvicorn
requests
pybreaker
prometheus-fastapi-instrumentator
prometheus-client
pydantic-settings